feat: update API to fetch data from actual DB instead of mock data
